# openapi.yaml
openapi: 3.0.3
info:
  title: Beacon API
  version: "1.0.0"
  description: |-
    Beacon is a crisis-damage crowdsourcing backend: anonymous mobile reporters
    submit building-damage reports (identified only by an `X-Device-Id` header),
    and authenticated analysts (JWT bearer) verify, dispatch and export them.

    ## Visibility tiers
    - **Anonymous / public** callers (no JWT) and the authenticated read-only
      `external_viewer` role see **verified reports only**, in a locked-down
      *public projection*: coordinates coarsened to 3 decimal places (~110 m),
      and `submitterId`, `what3words`, `plusCode`, `landmark`, `buildingId`,
      `buildingSource`, `infraName`, `gpsAccuracyMeters`, `description`,
      tasking/operational fields, `modular` and `anonymization` are stripped.
      Unverified reports return 404 so their existence and position are never
      leaked. Public aggregates (`/reports/area-groups`) count verified
      reports only for this tier, coherent with the verified-only public map.
    - **Analyst roles** (`field_validator`, `co_analyst`, `regional_analyst`,
      `crisis_admin`) see all statuses at full precision, limited to the
      crises in their `crisisScope` (`"*"` = org-wide).

    ## Crisis scoping
    Where `crisisId` is optional (stats, export), an omitted value resolves to
    the **newest active crisis** (the same crisis `GET /v1/crises/active`
    returns), falling back to a seeded default only when no crisis is active.

...

...

...
servers:
...
paths:
  /v1/auth/login:
    post:
      tags: [auth]
      operationId: login
      summary: Analyst login (email + password → JWT)
      description: |-
        Exchanges analyst credentials for a signed JWT and the user profile.
        Use the token as `Authorization: Bearer <token>` on analyst endpoints.
        Unknown email and wrong password return the identical 401 (no user
        enumeration).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          description: Authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          description: Invalid email or password (`error=invalid_credentials`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/auth/me:
    get:
      tags: [auth]
      operationId: me
      summary: Current authenticated analyst
      security:
        - bearerAuth: []
      responses:
        "200":
          description: The authenticated analyst account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "401":
          $ref: '#/components/responses/Unauthorized'
  /v1/reports:
    post:
      tags: [reports]
      operationId: submitReport
      summary: Submit a damage report (anonymous, idempotent)
      description: |-
        Anonymous mobile submission. The `X-Device-Id` header is **required**:
        it stamps the anonymous submitter (so "My Reports" works without
        de-anonymizing the reporter) and keys the per-device rate-limit and
        dedup guards. Requests without it are rejected with 400
        (`error=device_id_required`).

        **Idempotent re-submit:** `id` is client-supplied and is the
        idempotency anchor (`idempotencyKey` defaults to `idem-<id>`).
        Re-POSTing a report whose `id` or `idempotencyKey` is already stored
        inserts nothing and returns **200** with the stored report — an
        offline client can safely replay its queue. A first-time insert
        returns **201**. Replays never trip the rate-limit or dedup guards.

        **Location (nullable lat/lng):** a report has EITHER a resolved point
        (GPS fix or tapped building footprint) OR is location-unresolved
        (landmark-only). Send `lat`/`lng` as numbers with
        `locationResolved: true` (the default when omitted), or
        `lat: null, lng: null, locationResolved: false` plus a non-empty
        `landmark`. The server never stores or emits 0,0 (Null Island).
        `accuracyMeters` is the wire name for GPS fix accuracy; it coalesces
        with `gpsAccuracyMeters` and `location.gpsAccuracyMeters`.

        **Crisis assignment:** an explicit `crisisId` pins the report; empty
        or `"auto"` lets the server assign by space+time. With no match the
        report stays pending (empty `crisisId`) and may seed an emergent
        crisis proposal. **Geographic containment:** a pinned crisis is
        IGNORED (not rejected) when the report's resolved point lies more than
        max(radiusKm×1.5, radiusKm+5) km from that crisis's center — the
        normal spatial assignment runs instead. Landmark-only (no-coords)
        reports always keep their pin.

        **Hazard:** an empty `crisisNature` stays empty — the server never
        fabricates a default hazard.

        **Anti-abuse:** per submitter, at most 5 reports/minute and
        20 reports/10 minutes (429 `rate_limited`); a pin within 25 m and
        10 min of the same submitter's previous pin is rejected with 409
        `duplicate` referencing the existing report id. Only a real tapped
        footprint (`buildingSource: "footprint"`, or a legacy `fp-`-prefixed
        `buildingId`) is exempt — its re-reports merge via the per-building
        version chain; synthetic GPS-grid `b-` building ids are checked like
        building-less pins.

        **Modular secondary impacts** (`modular`): the server stores the blob
        verbatim; the documented wire enums are in the `ModularImpacts`
        schema (electricity, healthServices, pressingNeeds — incl.
        `basic_services` and the free-text `pressingNeedsOther`).
      parameters:
        - $ref: '#/components/parameters/DeviceIdRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitReportRequest'
      responses:
        "201":
          description: Report created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
        "200":
          description: Idempotent replay — the report already existed; the stored document is returned and nothing was inserted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
        "400":
          description: Malformed JSON (`bad_json`), missing `X-Device-Id` (`device_id_required`) or failed validation (`validation`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "409":
          description: Near-duplicate of an existing nearby report by the same submitter.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DuplicateError'
        "429":
          description: Per-device submit rate limit exceeded (`rate_limited`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      tags: [reports]
      operationId: listReports
      summary: List reports (analyst console; filter + paginate)
      description: |-
        Analyst-only list (any authenticated role; 401 without a token).

        **Scope:** an explicit `crisisId` must be inside the caller's
        `crisisScope` (403 `out_of_scope` otherwise). With no `crisisId`, an
        org-wide (`"*"`) analyst lists every crisis; a finitely-scoped analyst
        lists all crises in their scope.

        **Pagination:** offset (`page` × `pageSize`) or keyset via `cursor`
        (opaque base64url token from `nextCursor`; when supplied, `page` is
        ignored). Ordered by `capturedAt` DESC, `id` DESC. `pageSize` is
        capped at 200.

        The `external_viewer` role is forced to verified-only results in the
        coarsened public projection.
      security:
        - bearerAuth: []
      parameters:
        - name: crisisId
          in: query
          schema: { type: string }
          description: Restrict to one crisis (must be in the caller's scope).
        - name: damage
          in: query
          schema: { type: string }
          description: "Comma-separated or repeated damage grades (either vocabulary): none,slight,moderate,severe,destroyed,minimal,partial,complete."
        - name: verification
          in: query
          schema: { type: string }
          description: "Comma-separated or repeated: pending,verified,flagged."
        - name: q
          in: query
          schema: { type: string }
          description: Free-text — matches place (substring), id (prefix) or an infrastructure type.
        - name: mine
          in: query
          schema: { type: boolean }
          description: Only the caller's own reports; requires the `X-Device-Id` header (400 otherwise).
        - name: buildingId
          in: query
          schema: { type: string }
        - name: adm1Pcode
          in: query
          schema: { type: string }
          description: "Admin-1 code (geoBoundaries shapeID `GB:<id>` or seed code — not an OCHA P-code)."
        - name: adm2Pcode
          in: query
          schema: { type: string }
          description: Admin-2 code (same provenance caveat).
        - name: adm3Pcode
          in: query
          schema: { type: string }
          description: Admin-3 code (same provenance caveat).
        - name: taskStatus
          in: query
          schema: { type: string }
          description: "Comma-separated or repeated: new,triaged,assigned,in_progress,resolved,closed."
        - name: severity
          in: query
          schema: { type: string }
          description: "Comma-separated or repeated: routine,elevated,life_safety."
        - name: lifeSafety
          in: query
          schema: { type: boolean }
          description: Only life-safety reports.
        - name: assignee
          in: query
          schema: { type: string }
        - name: cluster
          in: query
          schema: { type: string }
          description: "One humanitarian cluster: slsc,health,wash,education,food_security,protection,logistics,nutrition,etc,cccm,early_recovery."
        - $ref: '#/components/parameters/BBox'
        - name: page
          in: query
          schema: { type: integer, default: 0, minimum: 0 }
          description: Zero-based page index (offset pagination; ignored when `cursor` is set).
        - name: pageSize
          in: query
          schema: { type: integer, default: 50, maximum: 200 }
        - name: cursor
          in: query
          schema: { type: string }
          description: Opaque keyset cursor from a previous response's `nextCursor`.
        - name: X-Device-Id
          in: header
          required: false
          schema: { type: string }
          description: Required only with `mine=true` (resolves the caller's anonymous submitter identity).
      responses:
        "200":
          description: One page of reports.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListResponse'
        "400":
          description: Bad bbox (`bad_bbox`), bad cursor (`bad_cursor`) or `mine=true` without `X-Device-Id` (`validation`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/OutOfScope'
  /v1/reports/export:
    get:
      tags: [exports]
      operationId: exportReports
      summary: Bulk export (GeoJSON / HXL CSV / GeoPackage / KML / PDNA pivot)
      description: |-
        Analyst-only bulk export with exact coordinates and the full
        per-report row. The read-only `external_viewer` role is **denied
        (403)** — its read need is served by the coarsened public map
        endpoints instead. `crisisId` defaults to the newest active crisis.

        **Row formats (geojson / csv / gpkg / kml)** carry the gate field
        list: `damage_classification` ∈ {Minimal, Partial, Complete} (the
        canonical 3-tier rollup), raw `damage` grade, `infrastructure_type`,
        `infrastructure_name`, `hazard_type` (empty when the reporter asserted
        none), `timestamp` (ISO-8601 of capturedAt, UTC), and the modular
        secondary-impact sections flattened to snake_case — the three known
        sections under their stable names (`electricity`, `health_services`,
        `pressing_needs`, always present) plus any later-added section
        DYNAMICALLY (e.g. `pressingNeedsOther` → `pressing_needs_other`,
        appended after the fixed CSV columns) — plus `possiblyDamaged`,
        `debris`, `buildingId`, `verification`, `place`, `description`
        (free-text note, analyst language), `plus_code`, `accuracy_m`, and
        the `admin1_shapeid` / `admin2_shapeid` / `admin3_shapeid` columns
        (geoBoundaries shapeIDs / seed codes — NOT OCHA P-codes; their HXL
        tags use `+id`, never `+code`).

        - `geojson` — FeatureCollection; geometry is a Point `[lng, lat]` in
          decimal degrees, or `null` for a location-unresolved report (never
          `[0,0]`).
        - `csv` — header row + an HXL hashtag row, `latitude`/`longitude`
          columns blank for unresolved reports.
        - `gpkg` — OGC GeoPackage (single SQLite file), NULL geometry for
          unresolved reports.
        - `kml` — one Placemark per *resolved* report (unresolved skipped).
        - `pdna` — a different shape: a PDNA-ready **damage-count** pivot
          (sector × admin-2) of report counts per grade, HXL-tagged CSV. The
          3-tier rollup (minimal/partial/complete) sums to `total`; the
          5-level EMS-98 columns are detail only. It is NOT a loss/cost
          estimation (no monetary figures).
      security:
        - bearerAuth: []
      parameters:
        - name: format
          in: query
          schema:
            type: string
            enum: [geojson, csv, gpkg, kml, pdna]
            default: geojson
        - name: crisisId
          in: query
          schema: { type: string }
          description: Defaults to the newest active crisis when omitted.
        - name: damage
          in: query
          schema: { type: string }
          description: Comma-separated or repeated damage grades.
        - name: verification
          in: query
          schema: { type: string }
          description: "Comma-separated or repeated: pending,verified,flagged."
        - name: q
          in: query
          schema: { type: string }
        - name: adm2Pcode
          in: query
          schema: { type: string }
        - name: adm3Pcode
          in: query
          schema: { type: string }
        - $ref: '#/components/parameters/BBox'
      responses:
        "200":
          description: The export file (Content-Disposition attachment).
          content:
            application/geo+json:
              schema:
                type: object
                description: GeoJSON FeatureCollection (format=geojson).
            text/csv:
              schema:
                type: string
                description: HXL-tagged CSV (format=csv) or PDNA damage-count pivot (format=pdna).
            application/geopackage+sqlite3:
              schema:
                type: string
                format: binary
                description: OGC GeoPackage (format=gpkg).
            application/vnd.google-earth.kml+xml:
              schema:
                type: string
                description: KML document (format=kml).
        "400":
          description: Unknown format (`bad_format`) or bad bbox (`bad_bbox`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          description: external_viewer cannot export raw data (`forbidden`), or the crisis is out of the caller's scope (`out_of_scope`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/reports/{id}:
    get:
      tags: [reports]
      operationId: getReport
      summary: Get one report (public = verified-only, coarsened; analyst = full)
      description: |-
        Public endpoint with tiered projection:
        - **Anonymous** (and the authenticated `external_viewer`): only a
          VERIFIED report is returned, in the public projection — coordinates
          coarsened to 3 decimals (~110 m); `submitterId`, `what3words`,
          `plusCode`, `landmark`, `buildingId`, `buildingSource`, `infraName`,
          `gpsAccuracyMeters`, `description`, tasking fields, `modular` and
          `anonymization` stripped. An unverified report 404s (its existence
          is never leaked).
        - **Analyst roles** in scope for the report's crisis get the full
          document at full precision (any verification status). An
          out-of-scope report 404s rather than 403s.
      parameters:
        - $ref: '#/components/parameters/ReportId'
      responses:
        "200":
          description: The report (full or public projection depending on caller tier).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
        "404":
          $ref: '#/components/responses/NotFound'
  /v1/reports/{id}/photo:
    post:
      tags: [photos]
      operationId: uploadReportPhoto
      summary: Upload the report's photo (anonymous, ownership-bound)
      description: |-
        Multipart upload (`file` part, max 12 MB). Only real JPEG/PNG/WEBP
        bytes are accepted (content-sniffed; 415 otherwise). The upload is
        bound to the report's creator: if the report stored its creating
        device, the caller's `X-Device-Id` must match (403 otherwise); if no
        device was recorded, only a FIRST upload (no photo yet) within 30
        minutes of capture is allowed.
      parameters:
        - $ref: '#/components/parameters/ReportId'
        - name: X-Device-Id
          in: header
          required: false
          schema: { type: string }
          description: Must match the report's creating device when one is recorded.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: JPEG, PNG or WEBP image (max 12 MB).
      responses:
        "200":
          description: Photo stored and recorded on the report.
          content:
            application/json:
              schema:
                type: object
                properties:
                  photoUrl:
                    type: string
                    example: /api/v1/reports/rep-123/photo
        "400":
          $ref: '#/components/responses/BadRequest'
        "403":
          description: Caller is not the reporting device, the photo already exists, or the first-upload window has closed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          $ref: '#/components/responses/NotFound'
        "415":
          description: The file is not a JPEG/PNG/WEBP image (`unsupported_media`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      tags: [photos]
      operationId: getReportPhoto
      summary: Serve the report's photo (visibility mirrors the report)
      description: |-
        Anonymous and `external_viewer` callers receive the photo only when
        the report is VERIFIED (404 otherwise — never reveals whether the
        file exists); `external_viewer` must also be in scope for the
        report's crisis. Analyst roles fetch any status within their crisis
        scope. Served with `Cache-Control: private`.
      parameters:
        - $ref: '#/components/parameters/ReportId'
      responses:
        "200":
          description: The image bytes.
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        "404":
          description: No such report/photo, or the caller's tier may not see it.
      security:
        - {}
        - bearerAuth: []
  /v1/reports/{id}/verification:
    patch:
      tags: [reports]
      operationId: patchVerification
      summary: Set a report's verification status (analyst decision)
      description: |-
        Mutator endpoint — any analyst role except the read-only
        `external_viewer` (403 `forbidden`), crisis-scoped: an analyst may
        only verify reports of crises in their scope (403 `out_of_scope`); a
        pending report with no crisis is mutable only with org-wide scope.

        **Photo gate:** setting `status: verified` on a report with NO photo
        is rejected with 409 `photo_required` unless `force: true` is passed.
        "Photo" means evidence actually **uploaded to the server** (a stored
        image, served by `GET /v1/reports/{id}/photo`) — photo metadata merely
        claimed by the client in the submit body does not count. The optional
        `note` and the `force` flag are recorded in the verification audit
        trail.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ReportId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerificationRequest'
      responses:
        "200":
          description: The updated report.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/OutOfScope'
        "404":
          $ref: '#/components/responses/NotFound'
        "409":
          description: |-
            Verifying a photo-less report without `force=true`
            (`photo_required`). "Photo-less" = no evidence uploaded to the
            server; client-claimed photo metadata does not satisfy the gate.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/reports/{id}/task:
    patch:
      tags: [reports]
      operationId: patchTask
      summary: Advance a report's dispatch task (analyst)
      description: |-
        Mutator endpoint (same auth/scope rules as the verification PATCH).
        Any subset of the body advances the task: status, assignee, severity,
        cluster routing, disposition, or an audit note.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ReportId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequest'
      responses:
        "200":
          description: The updated report.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/OutOfScope'
        "404":
          $ref: '#/components/responses/NotFound'
  /v1/buildings/{buildingId}/timeline:
    get:
      tags: [buildings]
      operationId: buildingTimeline
      summary: A building's damage version history
      description: |-
        Public endpoint. For the public tier (anonymous or `external_viewer`)
        it returns VERIFIED entries only, with each entry's free-text note
        stripped and the reporter shown as `community`. Analyst roles see the
        full chain (all statuses, notes, reporter).
      parameters:
        - name: buildingId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: The building's version chain.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuildingTimeline'
        "404":
          $ref: '#/components/responses/NotFound'
  /v1/stats/overview:
    get:
      tags: [stats]
      operationId: statsOverview
      summary: Aggregated crisis statistics (analyst)
      description: |-
        All aggregates computed in SQL for one crisis. `crisisId` defaults to
        the **newest active crisis** when omitted. The caller must be in
        scope for the resolved crisis (403 `out_of_scope`). For the
        `external_viewer` role the embedded `recent` list is reduced to
        verified reports in the public projection.
      security:
        - bearerAuth: []
      parameters:
        - name: crisisId
          in: query
          schema: { type: string }
          description: Defaults to the newest active crisis when omitted.
      responses:
        "200":
          description: The overview aggregates.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsOverview'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/OutOfScope'
  /v1/crises:
    get:
      tags: [crises]
      operationId: listCrises
      summary: List crises (optionally by status)
      parameters:
        - name: status
          in: query
          schema: { type: string }
          description: "Comma-separated lifecycle filter, e.g. `active,proposed`. Statuses: active, proposed, closed, dismissed."
      responses:
        "200":
          description: Crises.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Crisis'
  /v1/crises/active:
    get:
      tags: [crises]
      operationId: activeCrisis
      summary: The newest active crisis
      description: Returns the newest active crisis, or 204 when no crisis is active.
      responses:
        "200":
          description: The active crisis.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Crisis'
        "204":
          description: No active crisis.
  /v1/crises/near:
    get:
      tags: [crises]
      operationId: nearbyCrises
      summary: Active/proposed crises near a point (location-first launch)
      parameters:
        - name: lat
          in: query
          required: true
          schema: { type: number, minimum: -90, maximum: 90 }
        - name: lng
          in: query
          required: true
          schema: { type: number, minimum: -180, maximum: 180 }
        - name: radiusKm
          in: query
          schema: { type: number, default: 50 }
      responses:
        "200":
          description: Nearby crises with distance (and whether the point is covered).
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Crisis'
                  count:
                    type: integer
        "400":
          $ref: '#/components/responses/BadRequest'
  /v1/crises/{id}:
    get:
      tags: [crises]
      operationId: getCrisis
      summary: One crisis by id
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: The crisis.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Crisis'
        "404":
          $ref: '#/components/responses/NotFound'
  /v1/crises/{id}/form:
    patch:
      tags: [forms]
      operationId: patchCrisisForm
      summary: Adjust a crisis's modular capture form (senior analysts)
      description: |-
        Stores per-crisis overrides over the default Appendix-1 sections:
        section keys in `required` are flipped to required, keys in `disabled`
        are removed from the form entirely. Keys must name known sections
        (`electricity`, `healthServices`, `pressingNeeds`) and a key cannot be
        in both lists. Sending two empty lists resets the crisis to the pure
        defaults.

        **RBAC:** `regional_analyst` / `crisis_admin` only (shaping what every
        reporter in a crisis is asked is a senior decision, like
        `/crises/{id}/status`), and the crisis must be inside the caller's
        `crisisScope`. Other mutator roles get 403 `forbidden`; out-of-scope
        gets 403 `out_of_scope`.

        Responds with the crisis's RESOLVED schema (same shape as
        `GET /v1/form-schema`) so a dashboard can re-render immediately.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FormOverrides'
      responses:
        "200":
          description: The crisis's resolved form schema after the update.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FormSchema'
        "400":
          description: Malformed JSON (`bad_json`), unknown section key or a key in both lists (`validation`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/OutOfScope'
        "404":
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Analyst JWT issued by POST /v1/auth/login.
...
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
    LoginResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
          description: Signed JWT for the Authorization Bearer header.
        user:
          $ref: '#/components/schemas/User'
...